Sparse storage working, names changed
LOL YEEES! Sitting on the bus 177 just by Wallenbergs Graves, on my way to see Hampus and Forza. Maybe tomorrow go to Trånholmen. IDK why this didn't work, the hashtag (#) might have been messing with something, now store the colors in hex but without the #. Woop cheers. Remember to relax, this was hectic. / Jacob 21.23 on the 177, 22/7 2017.