Collision not working, never worked
The collision count counts like a ass, three bullets can lead to like 20 "collisions". Problem could be that when a player delete a bullet it might still be in an other players list of penetratinf bullets. This should only happen close to borders (where bullets disappear. This should also lead to collision count counting less hits than it should. Idk what is going on...